requirements.txt
numpy==1.23.4
pandas==1.5.3
polygon-api-client
PyMySQL==1.1.0
python_dateutil==2.8.2
Requests==2.31.0
SQLAlchemy==2.0.15
